Correctly reset deeplima between files
No more crash when analyzing several files

+++ deeplima/include/deeplima/eigen_wrp/word_seq_embd_vectorizer.h
@@ -532,7 +532,7 @@ public:
       m_pModel->precompute_inputs(input, Parent::m_precomputed_vectors[Parent::m_curr_bucket_id], 0);
       Parent::m_curr_bucket_id++;
 
-      if (Parent::m_curr_bucket_id >= Parent::m_precomputed_vectors.size())
+      if ((size_t)Parent::m_curr_bucket_id >= Parent::m_precomputed_vectors.size())
       {
         Parent::m_precomputed_vectors.resize(Parent::m_curr_bucket_id + 1);
       }
